% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIShinylab5.R
\name{koladaapi}
\alias{koladaapi}
\title{A function to get the URL of kolada.se KPIs}
\usage{
koladaapi()
}
\value{
a data frame of 3 columns of year,gender and value(salary)
}
\description{
A function to get the URL of kolada.se KPIs
}
